+++ lib/msal-core/src/UserAgentApplication.ts
@@ -384,7 +384,7 @@ export class UserAgentApplication {
           this._cacheStorage.setItem(authorityKey, this.authority);
         }
 
-        const urlNavigate = authenticationRequest.createNavigateUrl(scopes) + "&prompt=select_account" + "&response_mode=fragment";
+        const urlNavigate = authenticationRequest.createNavigateUrl(scopes)  + Constants.prompt_select_account + Constants.response_mode_fragment;
         this.promptUser(urlNavigate);
       });
   }
@@ -403,14 +403,14 @@ export class UserAgentApplication {
      */
     return new Promise<string>((resolve, reject) => {
       if (this._loginInProgress) {
-        reject(ErrorCodes.loginProgressError + "|" + ErrorDescription.loginProgressError);
+        reject(ErrorCodes.loginProgressError + Constants.resourceDelimeter + ErrorDescription.loginProgressError);
         return;
       }
 
       if (scopes) {
         const isValidScope = this.validateInputScope(scopes);
         if (isValidScope && !Utils.isEmpty(isValidScope)) {
-          reject(ErrorCodes.inputScopesError + "|" + ErrorDescription.inputScopesError);
+          reject(ErrorCodes.inputScopesError + Constants.resourceDelimeter + ErrorDescription.inputScopesError);
           return;
         }
 
@@ -434,9 +434,9 @@ export class UserAgentApplication {
           authenticationRequest.extraQueryParameters = extraQueryParameters;
         }
 
-        this._cacheStorage.setItem(Constants.loginRequest, window.location.href);
+        this._cacheStorage.setItem(Constants.loginRequest, window.location.href, this.storeAuthStateInCookie);
         this._cacheStorage.setItem(Constants.loginError, "");
-        this._cacheStorage.setItem(Constants.nonceIdToken, authenticationRequest.nonce);
+        this._cacheStorage.setItem(Constants.nonceIdToken, authenticationRequest.nonce, this.storeAuthStateInCookie);
         this._cacheStorage.setItem(Constants.msalError, "");
         this._cacheStorage.setItem(Constants.msalErrorDescription, "");
         const authorityKey = Constants.authority + Constants.resourceDelimeter + authenticationRequest.state;
@@ -444,7 +444,7 @@ export class UserAgentApplication {
           this._cacheStorage.setItem(authorityKey, this.authority);
         }
 
-        const urlNavigate = authenticationRequest.createNavigateUrl(scopes) + "&prompt=select_account" + "&response_mode=fragment";
+        const urlNavigate = authenticationRequest.createNavigateUrl(scopes)  + Constants.prompt_select_account  + Constants.response_mode_fragment;
         window.renewStates.push(authenticationRequest.state);
         window.requestType = Constants.login;
         this.registerCallback(authenticationRequest.state, scope, resolve, reject);
@@ -500,7 +500,7 @@ export class UserAgentApplication {
       this._cacheStorage.setItem(Constants.msalError, ErrorCodes.popUpWindowError);
       this._cacheStorage.setItem(Constants.msalErrorDescription, ErrorDescription.popUpWindowError);
       if (reject) {
-        reject(ErrorCodes.popUpWindowError + "|" + ErrorDescription.popUpWindowError);
+        reject(ErrorCodes.popUpWindowError + Constants.resourceDelimeter + ErrorDescription.popUpWindowError);
       }
       return null;
     }
@@ -509,11 +509,11 @@ export class UserAgentApplication {
     var pollTimer = window.setInterval(() => {
       if (popupWindow && popupWindow.closed && instance._loginInProgress) {
         if (reject) {
-          reject(ErrorCodes.userCancelledError + "|" + ErrorDescription.userCancelledError);
+          reject(ErrorCodes.userCancelledError + Constants.resourceDelimeter + ErrorDescription.userCancelledError);
         }
         window.clearInterval(pollTimer);
         if (this._isAngular) {
-            this.broadcast('msal:popUpClosed', ErrorCodes.userCancelledError + "|" + ErrorDescription.userCancelledError);
+            this.broadcast('msal:popUpClosed', ErrorCodes.userCancelledError + Constants.resourceDelimeter + ErrorDescription.userCancelledError);
             return;
         }
         instance._loginInProgress = false;
@@ -684,7 +684,7 @@ export class UserAgentApplication {
           for (let i = 0; i < window.callBacksMappedToRenewStates[expectedState].length; ++i) {
             try {
               if (errorDesc || error) {
-                  window.callBacksMappedToRenewStates[expectedState][i].reject(errorDesc + "|" + error);
+                  window.callBacksMappedToRenewStates[expectedState][i].reject(errorDesc + Constants.resourceDelimeter + error);
               }
               else if (token) {
                   window.callBacksMappedToRenewStates[expectedState][i].resolve(token);
@@ -891,35 +891,45 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
    * @ignore
    * @hidden
    */
-  private addHintParameters(urlNavigate: string, user: User): string {
-    const userObject = user ? user : this.getUser();
-    const decodedClientInfo = userObject.userIdentifier.split(".");
-    const uid = Utils.base64DecodeStringUrlSafe(decodedClientInfo[0]);
-    const utid = Utils.base64DecodeStringUrlSafe(decodedClientInfo[1]);
-    if (!this.urlContainsQueryStringParameter("login_hint", urlNavigate) && userObject.displayableId && !Utils.isEmpty(userObject.displayableId)) {
-      urlNavigate += "&login_hint=" + encodeURIComponent(user.displayableId);
-    }
+    private addHintParameters(urlNavigate: string, user: User): string {
+        const userObject = user ? user : this.getUser();
+        if (userObject) {
+            const decodedClientInfo = userObject.userIdentifier.split(".");
+            const uid = Utils.base64DecodeStringUrlSafe(decodedClientInfo[0]);
+            const utid = Utils.base64DecodeStringUrlSafe(decodedClientInfo[1]);
 
-    if (!Utils.isEmpty(uid) && !Utils.isEmpty(utid)) {
-      if (!this.urlContainsQueryStringParameter("domain_req", urlNavigate) && !Utils.isEmpty(utid)) {
-        urlNavigate += "&domain_req=" + encodeURIComponent(utid);
-      }
+            if (userObject.sid  && urlNavigate.indexOf(Constants.prompt_none) !== -1) {
+                if (!this.urlContainsQueryStringParameter(Constants.sid, urlNavigate) && !this.urlContainsQueryStringParameter(Constants.login_hint, urlNavigate)) {
+                    urlNavigate += "&" + Constants.sid +"=" + encodeURIComponent(this._user.sid);
+                }
+            }
+            else {
+                if (!this.urlContainsQueryStringParameter(Constants.login_hint, urlNavigate) && userObject.displayableId && !Utils.isEmpty(userObject.displayableId)) {
+                    urlNavigate += "&" + Constants.login_hint +"=" + encodeURIComponent(user.displayableId);
+                }
+            }
 
-      if (!this.urlContainsQueryStringParameter("login_req", urlNavigate) && !Utils.isEmpty(uid)) {
-        urlNavigate += "&login_req=" + encodeURIComponent(uid);
-      }
+            if (!Utils.isEmpty(uid) && !Utils.isEmpty(utid)) {
+                if (!this.urlContainsQueryStringParameter("domain_req", urlNavigate) && !Utils.isEmpty(utid)) {
+                    urlNavigate += "&domain_req=" + encodeURIComponent(utid);
+                }
 
-      if (!this.urlContainsQueryStringParameter("domain_hint", urlNavigate) && !Utils.isEmpty(utid)) {
-        if (utid === "9188040d-6c67-4c5b-b112-36a304b66dad") {
-          urlNavigate += "&domain_hint=" + encodeURIComponent("consumers");
-        } else {
-          urlNavigate += "&domain_hint=" + encodeURIComponent("organizations");
+                if (!this.urlContainsQueryStringParameter("login_req", urlNavigate) && !Utils.isEmpty(uid)) {
+                    urlNavigate += "&login_req=" + encodeURIComponent(uid);
+                }
+            }
+            if (!this.urlContainsQueryStringParameter(Constants.domain_hint, urlNavigate) && !Utils.isEmpty(utid)) {
+                if (utid === "9188040d-6c67-4c5b-b112-36a304b66dad") {
+                    urlNavigate += "&" +  Constants.domain_hint + "=" + encodeURIComponent("consumers");
+                } else {
+                    urlNavigate += "&" + Constants.domain_hint + "=" + encodeURIComponent("organizations");
+                }
+            }
+
         }
-      }
-    }
 
-    return urlNavigate;
-  }
+        return urlNavigate;
+    }
 
   /*
    * Checks if the authorization endpoint URL contains query string parameters
@@ -966,12 +976,13 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
     }
 
     const scope = scopes.join(" ").toLowerCase();
-    if (!userObject) {
-      if (this._tokenReceivedCallback) {
-        this._tokenReceivedCallback(ErrorDescription.userLoginError, null, ErrorCodes.userLoginError, Constants.accessToken, this.getUserState(this._cacheStorage.getItem(Constants.stateLogin)));
-        return;
+      if (!userObject && !(extraQueryParameters && (extraQueryParameters.indexOf(Constants.login_hint) !== -1 ))) {
+          if (this._tokenReceivedCallback) {
+              this._logger.info('User login is required');
+              this._tokenReceivedCallback(ErrorDescription.userLoginError, null, ErrorCodes.userLoginError, Constants.accessToken, this.getUserState(this._cacheStorage.getItem(Constants.stateLogin)));
+              return;
+          }
       }
-    }
 
     this._acquireTokenInProgress = true;
     let authenticationRequest: AuthenticationRequestParameters;
@@ -989,8 +1000,15 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
         authenticationRequest = new AuthenticationRequestParameters(acquireTokenAuthority, this.clientId, scopes, ResponseTypes.id_token_token, this._redirectUri, this._state);
       }
 
-      this._cacheStorage.setItem(Constants.nonceIdToken, authenticationRequest.nonce);
-      const acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + userObject.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
+      this._cacheStorage.setItem(Constants.nonceIdToken, authenticationRequest.nonce, this.storeAuthStateInCookie);
+      var acquireTokenUserKey;
+      if(userObject) {
+           acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + userObject.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
+      }
+      else {
+          acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter  + Constants.no_user +Constants.resourceDelimeter +  authenticationRequest.state;
+      }
+
       if (Utils.isEmpty(this._cacheStorage.getItem(acquireTokenUserKey))) {
         this._cacheStorage.setItem(acquireTokenUserKey, JSON.stringify(userObject));
       }
@@ -1004,10 +1022,10 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
         authenticationRequest.extraQueryParameters = extraQueryParameters;
       }
 
-      let urlNavigate = authenticationRequest.createNavigateUrl(scopes) + "&prompt=select_account" + "&response_mode=fragment";
+      let urlNavigate = authenticationRequest.createNavigateUrl(scopes) +  Constants.prompt_select_account  + Constants.response_mode_fragment;
       urlNavigate = this.addHintParameters(urlNavigate, userObject);
       if (urlNavigate) {
-        this._cacheStorage.setItem(Constants.stateAcquireToken, authenticationRequest.state);
+        this._cacheStorage.setItem(Constants.stateAcquireToken, authenticationRequest.state, this.storeAuthStateInCookie);
         window.location.replace(urlNavigate);
       }
     });
@@ -1033,7 +1051,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
     return new Promise<string>((resolve, reject) => {
       const isValidScope = this.validateInputScope(scopes);
       if (isValidScope && !Utils.isEmpty(isValidScope)) {
-        reject(ErrorCodes.inputScopesError + "|" + isValidScope);
+        reject(ErrorCodes.inputScopesError + Constants.resourceDelimeter + isValidScope);
       }
 
       if (scopes) {
@@ -1042,15 +1060,17 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
 
       const userObject = user ? user : this.getUser();
       if (this._acquireTokenInProgress) {
-        reject(ErrorCodes.acquireTokenProgressError + "|" + ErrorDescription.acquireTokenProgressError);
+        reject(ErrorCodes.acquireTokenProgressError + Constants.resourceDelimeter + ErrorDescription.acquireTokenProgressError);
         return;
       }
 
       const scope = scopes.join(" ").toLowerCase();
-      if (!userObject) {
-        reject(ErrorCodes.userLoginError + "|" + ErrorDescription.userLoginError);
-        return;
-      }
+        //if user is not currently logged in and no login_hint is passed
+        if (!userObject && !(extraQueryParameters && (extraQueryParameters.indexOf(Constants.login_hint) !== -1))) {
+            this._logger.info('User login is required');
+            reject(ErrorCodes.userLoginError + Constants.resourceDelimeter + ErrorDescription.userLoginError);
+            return;
+        }
 
       this._acquireTokenInProgress = true;
       let authenticationRequest: AuthenticationRequestParameters;
@@ -1074,8 +1094,15 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
 
         this._cacheStorage.setItem(Constants.nonceIdToken, authenticationRequest.nonce);
         authenticationRequest.state = authenticationRequest.state;
-        const acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + userObject.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
-        if (Utils.isEmpty(this._cacheStorage.getItem(acquireTokenUserKey))) {
+        var acquireTokenUserKey;
+        if(userObject) {
+            acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + userObject.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
+        }
+         else {
+            acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter  + Constants.no_user +Constants.resourceDelimeter +  authenticationRequest.state;
+        }
+
+          if (Utils.isEmpty(this._cacheStorage.getItem(acquireTokenUserKey))) {
           this._cacheStorage.setItem(acquireTokenUserKey, JSON.stringify(userObject));
         }
 
@@ -1088,7 +1115,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
           authenticationRequest.extraQueryParameters = extraQueryParameters;
         }
 
-        let urlNavigate = authenticationRequest.createNavigateUrl(scopes) + "&prompt=select_account" + "&response_mode=fragment";
+        let urlNavigate = authenticationRequest.createNavigateUrl(scopes) + Constants.prompt_select_account  + Constants.response_mode_fragment;
         urlNavigate = this.addHintParameters(urlNavigate, userObject);
         window.renewStates.push(authenticationRequest.state);
         window.requestType = Constants.renewToken;
@@ -1102,7 +1129,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
         this._cacheStorage.setItem(Constants.msalError, ErrorCodes.endpointResolutionError);
         this._cacheStorage.setItem(Constants.msalErrorDescription, ErrorDescription.endpointResolutionError);
         if (reject) {
-          reject(ErrorCodes.endpointResolutionError + "|" + ErrorDescription.endpointResolutionError);
+          reject(ErrorCodes.endpointResolutionError + Constants.resourceDelimeter + ErrorDescription.endpointResolutionError);
         }
         if (popUpWindow) {
             popUpWindow.close();
@@ -1140,10 +1167,11 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
 
         const scope = scopes.join(" ").toLowerCase();
         const userObject = user ? user : this.getUser();
-        if (!userObject) {
-          reject(ErrorCodes.userLoginError + "|" + ErrorDescription.userLoginError);
-          return;
-        }
+          if (!userObject && !(extraQueryParameters &&  ((extraQueryParameters.indexOf(Constants.login_hint) !== -1 ||  extraQueryParameters.indexOf(Constants.sid) !== -1 )))) {
+              this._logger.info('User login is required');
+              reject(ErrorCodes.userLoginError + Constants.resourceDelimeter + ErrorDescription.userLoginError);
+              return;
+          }
 
         let authenticationRequest: AuthenticationRequestParameters;
         let newAuthority = authority ? AuthorityFactory.CreateInstance(authority, this.validateAuthority) : this.authorityInstance;
@@ -1167,7 +1195,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
           }
           else if (cacheResult.errorDesc || cacheResult.error) {
             this._logger.infoPii(cacheResult.errorDesc + ":" + cacheResult.error);
-            reject(cacheResult.errorDesc + "|" + cacheResult.error);
+            reject(cacheResult.errorDesc + Constants.resourceDelimeter + cacheResult.error);
             return;
           }
         }
@@ -1296,7 +1324,14 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
       authenticationRequest.extraQueryParameters = extraQueryParameters;
     }
 
-    const acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + user.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
+    var acquireTokenUserKey;
+    if(user) {
+        acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + user.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
+    }
+    else {
+        acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter  + Constants.no_user +Constants.resourceDelimeter +  authenticationRequest.state;
+    }
+
     if (Utils.isEmpty(this._cacheStorage.getItem(acquireTokenUserKey))) {
       this._cacheStorage.setItem(acquireTokenUserKey, JSON.stringify(user));
     }
@@ -1309,7 +1344,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
     // renew happens in iframe, so it keeps javascript context
     this._cacheStorage.setItem(Constants.nonceIdToken, authenticationRequest.nonce);
     this._logger.verbose("Renew token Expected state: " + authenticationRequest.state);
-    let urlNavigate = authenticationRequest.createNavigateUrl(scopes) + "&prompt=none";
+    let urlNavigate = authenticationRequest.createNavigateUrl(scopes) + Constants.prompt_none;
     urlNavigate = this.addHintParameters(urlNavigate, user);
     window.renewStates.push(authenticationRequest.state);
     window.requestType = Constants.renewToken;
@@ -1332,7 +1367,13 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
       authenticationRequest.extraQueryParameters = extraQueryParameters;
     }
 
-    const acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + user.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
+    var acquireTokenUserKey;
+    if(user) {
+        acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + user.userIdentifier + Constants.resourceDelimeter + authenticationRequest.state;
+    }
+    else {
+        acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + Constants.no_user + Constants.resourceDelimeter + authenticationRequest.state;
+    }
     if (Utils.isEmpty(this._cacheStorage.getItem(acquireTokenUserKey))) {
       this._cacheStorage.setItem(acquireTokenUserKey, JSON.stringify(user));
     }
@@ -1344,7 +1385,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
 
     this._cacheStorage.setItem(Constants.nonceIdToken, authenticationRequest.nonce);
     this._logger.verbose("Renew Idtoken Expected state: " + authenticationRequest.state);
-    let urlNavigate = authenticationRequest.createNavigateUrl(scopes) + "&prompt=none";
+    let urlNavigate = authenticationRequest.createNavigateUrl(scopes) + Constants.prompt_none;
     urlNavigate = this.addHintParameters(urlNavigate, user);
     window.renewStates.push(authenticationRequest.state);
     window.requestType = Constants.renewToken;
@@ -1455,7 +1496,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
   
     try {
         if (tokenReceivedCallback) {
-            tokenReceivedCallback.call(self, errorDesc, token, error, tokenType);
+            tokenReceivedCallback.call(self, errorDesc, token, error, tokenType,  this.getUserState(this._cacheStorage.getItem(Constants.stateLogin)));
         }
 
     } catch (err) {
@@ -1581,6 +1622,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
           }
 
           acquireTokenUserKey = Constants.acquireTokenUser + Constants.resourceDelimeter + user.userIdentifier + Constants.resourceDelimeter + tokenResponse.stateResponse;
+          var   acquireTokenUserKey_nouser = Constants.acquireTokenUser + Constants.resourceDelimeter  + Constants.no_user +Constants.resourceDelimeter +  tokenResponse.stateResponse;
           let acquireTokenUser: User;
           if (!Utils.isEmpty(this._cacheStorage.getItem(acquireTokenUserKey))) {
             acquireTokenUser = JSON.parse(this._cacheStorage.getItem(acquireTokenUserKey));
@@ -1593,10 +1635,12 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
                 "The user object created from the response is not the same as the one passed in the acquireToken request");
             }
           }
+          else if (!Utils.isEmpty(this._cacheStorage.getItem(acquireTokenUserKey_nouser))) {
+                  this.saveAccessToken(authority, tokenResponse, user, clientInfo, idToken);
+          }
         }
 
         if (tokenResponse.parameters.hasOwnProperty(Constants.idToken)) {
-          if (scope.indexOf(this.clientId) > -1) {
             this._logger.info("Fragment has id token");
             this._loginInProgress = false;
             idToken = new IdToken(tokenResponse.parameters[Constants.idToken]);
@@ -1635,12 +1679,11 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
               this._cacheStorage.setItem(Constants.msalError, "invalid idToken");
               this._cacheStorage.setItem(Constants.msalErrorDescription, "Invalid idToken. idToken: " + tokenResponse.parameters[Constants.idToken]);
             }
-          }
         }
       } else {
         authorityKey = tokenResponse.stateResponse;
         acquireTokenUserKey = tokenResponse.stateResponse;
-        this._logger.error("State Mismatch.Expected State: " + this._cacheStorage.getItem(Constants.stateLogin) + "," + "Actual State: " + tokenResponse.stateResponse);
+        this._logger.error("State Mismatch.Expected State: " + this._cacheStorage.getItem(Constants.stateLogin,this.storeAuthStateInCookie ) + "," + "Actual State: " + tokenResponse.stateResponse);
         tokenResponse.parameters['error'] = 'Invalid_state';
         tokenResponse.parameters['error_description'] = 'Invalid_state. state: ' + tokenResponse.stateResponse;
         this._cacheStorage.setItem(Constants.msalError, "Invalid_state");
@@ -1649,6 +1692,8 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
       }
       this._cacheStorage.setItem(Constants.renewStatus + tokenResponse.stateResponse, Constants.tokenRenewStatusCompleted);
       this._cacheStorage.removeAcquireTokenEntries(authorityKey, acquireTokenUserKey);
+      //this is required if navigateToLoginRequestUrl=false
+      this._cacheStorage.clearCookie();
   }
 
   /*
@@ -1717,7 +1762,7 @@ protected getCachedTokenInternal(scopes : Array<string> , user: User): CacheResu
           tokenResponse.requestType = Constants.login;
           tokenResponse.stateMatch = true;
           return tokenResponse;
-        } else if (stateResponse === this._cacheStorage.getItem(Constants.stateAcquireToken)) { //acquireTokenRedirect
+        } else if (stateResponse === this._cacheStorage.getItem(Constants.stateAcquireToken, this.storeAuthStateInCookie)) { //acquireTokenRedirect
           tokenResponse.requestType = Constants.renewToken;
           tokenResponse.stateMatch = true;
           return tokenResponse;